仮データを出力できるように調整した。
